requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==4.1.0
celery==5.3.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.7
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
kombu==5.3.2
langdetect==1.0.9
Markdown==3.4.3
multidict==6.0.4
openai==0.27.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
